Move logging into separate file
