+++ src/hash.cpp
@@ -20,6 +20,7 @@
 #include "hash.hpp"
 
 #include "legacy_util.hpp"
+#include "logging.hpp"
 
 #include <blake2.h>
